user authorization

+++ backend/routes/posts.js
@@ -39,8 +39,11 @@ router.post("",
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      //Id of user passed from check-auth with response from checkAuth
+      creator: req.userData.userId
     });
+
     post.save().then(createdPost => {
       res.status(201).json({
         message: 'Post added successfully',
@@ -69,8 +72,13 @@ multer({storage: storage}).single("image"), (req, res, next) => {
     imagePath: imagePath
   });
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: 'Update successful!'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0){
+      res.status(200).json({message: 'Update successful!'});
+    }else{
+      res.status(401).json({message: 'Not authorized!'});
+    }
+
   });
 });
 
@@ -115,9 +123,13 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then(
     result => {
-      res.status(200).json({message: 'Post deleted!'});
+      if (result.n > 0){
+        res.status(200).json({message: 'Post deleted!'});
+      }else{
+        res.status(401).json({message: 'Not authorized!'});
+      }
     }
   )
 });
